Adding a Scene to Hold Noesis Views
- From the Noesis people directly:
```
If you are programmatically assigning assets to NoesisView make sure those
assets are hard-references by Unity, by having them in the chain of dependencies
of the scene. For example, adding each one to a public propert of a
MonoBehaviour.
We have plans to better report these issues, when a XAML not correctly
referenced is being loaded.
```